move where QuixTopicManager lives (to be alongside TopicManager)
