+++ examples/bank_example/quix_platform_version/producer.py
@@ -5,12 +5,13 @@ from time import sleep
 
 from dotenv import load_dotenv
 
+from quixstreams import TopicManager
 from quixstreams.kafka import Producer
 from quixstreams.models.serializers import (
     QuixTimeseriesSerializer,
     SerializationContext,
 )
-from quixstreams.platforms.quix import QuixTopicManager, QuixKafkaConfigsBuilder
+from quixstreams.platforms.quix import QuixKafkaConfigsBuilder
 
 load_dotenv("./bank_example/quix_platform_version/quix_vars.env")
 
@@ -18,7 +19,7 @@ load_dotenv("./bank_example/quix_platform_version/quix_vars.env")
 # For non-"Application.Quix" platform producing, config is a bit manual right now
 cfg_builder = QuixKafkaConfigsBuilder()
 cfgs = cfg_builder.get_confluent_broker_config()
-topic_manager = QuixTopicManager(quix_config_builder=cfg_builder)
+topic_manager = TopicManager.Quix(quix_config_builder=cfg_builder)
 topic = topic_manager.topic("qts__purchase_events").name
 topic_manager.create_all_topics()
 serialize = QuixTimeseriesSerializer()
