+++ quixstreams/platforms/quix/__init__.py
@@ -1,4 +1,3 @@
 from .api import *
 from .config import *
 from .checks import *
-from .topic_manager import *
